Fix Stream#content
Get rid of the `eval`. I'm still not certain whether this should stay as JSON or have its own model. Without an object/model it will be harder to tie the UI elements that manipulate the filters back to something, and I'm concerned the JSON will eventually morph into a giant specification I have to keep track of.