Rip out the JS and rethink my approach
Also go back to two separate queries for Thoughts and Users when loading a Stream for display. The one giant query was weird in that it indivdually selected and named every field in both tables and did a left join on them, which makes no sense since it could obviously be an inner join. This way we avoid a join entirely and just directly query for the specific Users we need to load